Fixed disconnect with many enemies and editor redo/undo bug

+++ core/src/io/anuke/mindustry/world/Block.java
@@ -186,7 +186,7 @@ public class Block{
 			i++;
 			i %= 4;
 		}
-		tile.setDump((byte)pdump);
+		tile.setDump(pdump);
 		handleItem(item, tile, tile);
 	}
